Renamed Certificates.js -> CertificateCreator.js.
